requirements.txt
boto==2.49.0
boto3==1.9.14
certifi==2018.8.24
chardet==3.0.4
coverage==4.0.3
dj-database-url==0.5.0
Django==2.1.2
django-bootstrap-form==3.4
django-cors-headers==2.4.0
django-getenv==1.3.2
django-localflavor==2.1
django-oauth-toolkit==1.2.0
django-ses==0.8.6
django-settings-export==1.2.1
flake8==3.5.0
future==0.16.0
idna==2.7
mccabe==0.6.1
oauthlib==2.1.0
pluggy==0.7.1
psycopg2==2.7.5
py==1.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
python-coveralls==2.9.1
pytz==2018.5
PyYAML==4.2b1
requests==2.20.0
six==1.11.0
social-auth-app-django==2.1.0
git+https://github.com/mpdavis/python-jose.git@414e71b77718850e0208fa9f8682867676592dd9#egg=python-jose
pyjwkest==1.4.0
tox==3.2.1
urllib3==1.23
virtualenv==16.0.0
djangorestframework==3.9.1
djangorestframework-xml==1.4.0
